handle sv_password changes detected via log listener
Note: password is still checked regularly via the server verifier
timer. Eventually, this change will allow us to remove that. But for now
that verifier job also serves the purpose of making sure that the server
is still alive, so it has been left in tact until a better solution for
server liveliness checks is implemented.